backend/requirements.txt
open -e ~/Downloads/codemind/backend/requirements.txt
```

**Select all text** (`Cmd + A`), **delete it all**, then **paste this entire new version:**
```
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.1
pydantic-settings>=2.2.1

# LangChain ecosystem
langchain>=0.2.1
langchain-community>=0.2.1
langchain-openai>=0.1.8
langchain-core>=0.2.1
langgraph>=0.0.69

# Gemini (FREE)
langchain-google-genai>=1.0.6
google-generativeai>=0.7.2

# Vector DB
faiss-cpu>=1.9.0

# GitHub ingestion
GitPython>=3.1.43
PyGithub>=2.3.0
httpx>=0.27.0

# Embeddings & LLM
openai>=1.30.1
tiktoken>=0.7.0

# Utilities
python-dotenv>=1.0.1
aiofiles>=23.2.1
structlog>=24.2.0
tenacity>=8.3.0